Third commit
Everything working and tested. Missing webdev and instructions to the user, as well as setting up sql.

+++ tp3/controller/ControllerGenre.php
@@ -14,49 +14,63 @@ class ControllerGenre extends Controller {
     }
 
     public function create() {
-        return Twig::render('genre-create.php');
+        CheckSession::sessionAuth();
+        if($_SESSION['privilege'] == 1) {
+            return Twig::render('genre-create.php');
+        } else {
+            RequirePage::url('genre');
+        }
     }
 
     public function store() {  
-        $validation = new Validation;
-        extract($_POST);
-        $validation->name('nom_genre')->value($nom_genre)->max(20)->min(3);
+        if (isset($_POST) && !empty($_POST)) {
+            $validation = new Validation;
+            extract($_POST);
+            $validation->name('nom_genre')->value($nom_genre)->max(20)->min(3);
 
-        if(!$validation->isSuccess()) {
-            $errors = $validation->displayErrors();
+            if(!$validation->isSuccess()) {
+                $errors = $validation->displayErrors();
 
-            // return Twig::render('genre-create.php', ['errors' => $errors, 'genre' => $_POST]);
-            return Twig::render('genre-create.php', ['errors' => $errors]);
-            exit();
+                // return Twig::render('genre-create.php', ['errors' => $errors, 'genre' => $_POST]);
+                return Twig::render('genre-create.php', ['errors' => $errors]);
+                exit();
+            } else {
+                $genre = new Genre;
+                $insert = $genre->insert($_POST);
+                RequirePage::url('genre');
+            }
         } else {
-            $genre = new Genre;
-            $insert = $genre->insert($_POST);
             RequirePage::url('genre');
         }
     }
 
     public function destroy($id) {
-        $artiste = new Artiste;
-        $selectArtiste = $artiste->select('id_usager');
-        $suppressionAutorisee = true;
+        CheckSession::sessionAuth();
+        if ($_SESSION['privilege'] == 1) {
+            // Assurer qu'aucun genre lié à un utilisateur ne soit supprimé.
+            $artiste = new Artiste;
+            $selectArtiste = $artiste->select('id_usager');
+            $suppressionAutorisee = true;
 
-        foreach ($selectArtiste as $artiste) {
-            echo $artiste['id_genre'];
-            if ($id == $artiste['id_genre']) {
-                $suppressionAutorisee = false;
-                break;
+            foreach ($selectArtiste as $artiste) {
+                echo $artiste['id_genre'];
+                if ($id == $artiste['id_genre']) {
+                    $suppressionAutorisee = false;
+                    break;
+                }
             }
-        }
-        if ($suppressionAutorisee) {
-            $genre = new Genre;
-            $delete = $genre->delete($id);
+            if ($suppressionAutorisee) {
+                $genre = new Genre;
+                $delete = $genre->delete($id);
 
-            RequirePage::url('genre');
+                RequirePage::url('genre');
+            } else {
+                RequirePage::url('genre');
+            }   
         } else {
             RequirePage::url('genre');
-        }         
+        }     
     }
-
 }
 
 ?>
